+++ src/Checker/ResponseBuilder.php
@@ -4,22 +4,31 @@ declare(strict_types=1);
 
 namespace Umanit\LifePageBundle\Checker;
 
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\Response;
+use Umanit\LifePageBundle\Event\UnavailableServicesEvent;
 
 class ResponseBuilder implements ResponseBuilderInterface
 {
     /** @var ServiceCheckerCollectionInterface */
     private $checkersCollection;
 
-    public function __construct(ServiceCheckerCollectionInterface $checkersCollection)
-    {
+    /** @var EventDispatcherInterface */
+    private $eventDispatcher;
+
+    public function __construct(
+        ServiceCheckerCollectionInterface $checkersCollection,
+        EventDispatcherInterface $eventDispatcher
+    ) {
         $this->checkersCollection = $checkersCollection;
+        $this->eventDispatcher = $eventDispatcher;
     }
 
     public function buildResponse(string $type): Response
     {
         $response = '';
         $allOk = 1;
+        $unavailableServices = [];
 
         foreach ($this->checkersCollection->getCheckers($type)->checkAll() as $check) {
             $status = $check->getStatus();
@@ -27,8 +36,18 @@ class ResponseBuilder implements ResponseBuilderInterface
                 continue;
             }
 
-            $response .= sprintf('%s: %s', $check->getName(), $check->getTextStatus()).PHP_EOL;
+            $response .= sprintf('%s: %s', $check->getName(), $check->getTextStatus()) . PHP_EOL;
             $allOk &= (int) $status;
+
+            if (false === $status) {
+                $unavailableServices[] = $check->getName();
+            }
+        }
+
+        if (!empty($unavailableServices)) {
+            $event = new UnavailableServicesEvent($unavailableServices);
+
+            $this->eventDispatcher->dispatch($event);
         }
 
         return new Response(